添加shiro

+++ src/main/java/com/exposure/exposureservice/controller/admin/SysRoleController.java
@@ -41,7 +41,7 @@ public class SysRoleController {
 
     @ApiOperation(value = "findById", notes = "根据ID查询系统角色详情", response = SysRole.class)
     @GetMapping("/findById")
-    public ResultBean<Object> findById(@RequestParam("id") Long id) {
+    public ResultBean<Object> findById(@RequestParam("id") Integer id) {
         SysRole sysRole = sysRoleService.findById(id);
         return ResultUtils.success(sysRole);
     }
@@ -62,8 +62,8 @@ public class SysRoleController {
 
     @ApiOperation(value = "del", notes = "删除系统角色")
     @PostMapping("/del")
-    public ResultBean<Object> del(@RequestBody @ApiParam(name = "id", value = "id", required = true) Map<String, Long> map) {
-        Long id = map.get("id");
+    public ResultBean<Object> del(@RequestBody @ApiParam(name = "id", value = "id", required = true) Map<String, Integer> map) {
+        Integer id = map.get("id");
         sysRoleService.del(id);
         return ResultUtils.success();
     }
